Move methods from Arrays to more appropriate modules
