+++ lib/ripper_ruby_parser/sexp_handlers.rb
@@ -2,7 +2,6 @@
 
 require 'ripper_ruby_parser/sexp_handlers/helper_methods'
 
-require 'ripper_ruby_parser/sexp_handlers/arrays'
 require 'ripper_ruby_parser/sexp_handlers/assignment'
 require 'ripper_ruby_parser/sexp_handlers/blocks'
 require 'ripper_ruby_parser/sexp_handlers/conditionals'
@@ -22,7 +21,6 @@ module RipperRubyParser
       base.class_eval do
         include HelperMethods
 
-        include Arrays
         include Assignment
         include Blocks
         include Conditionals
