+++ test/ripper_ruby_parser/sexp_handlers/method_calls_test.rb
@@ -139,6 +139,26 @@ describe RipperRubyParser::Parser do
         end
       end
 
+      describe 'for collection indexing' do
+        it 'works in the simple case' do
+          'foo[bar]'.
+            must_be_parsed_as s(:call,
+                                s(:call, nil, :foo),
+                                :[],
+                                s(:call, nil, :bar))
+        end
+
+        it 'works without any indexes' do
+          'foo[]'.must_be_parsed_as s(:call, s(:call, nil, :foo),
+                                      :[])
+        end
+
+        it 'works with self[]' do
+          'self[foo]'.must_be_parsed_as s(:call, s(:self), :[],
+                                          s(:call, nil, :foo))
+        end
+      end
+
       describe 'safe call' do
         it 'works without arguments' do
           'foo&.bar'.must_be_parsed_as s(:safe_call, s(:call, nil, :foo), :bar)
